+++ stream-chat-android-core/src/main/java/io/getstream/chat/android/models/Member.kt
@@ -101,7 +101,7 @@ public data class Member(
             "channel_role", "channelRole" -> channelRole
             "notifications_muted", "notificationsMuted" -> notificationsMuted
             "status" -> status
-            else -> null
+            else -> extraData[fieldName] as? Comparable<*>
         }
     }
 }